[AD-316] Split DocumentDbStatementBasicTest.java adding DocumentDbStatementDateTimeTest.java https://bitquill.atlassian.net/browse/AD-316
